move networks config from utils package to config package
