+++ cmd/devtools/config/networks.go
@@ -1,4 +1,4 @@
-package utilities
+package config
 
 import (
 	"encoding/json"
